Make writing to tempfile and execute more generic
